refactor: rename module in api and add a Payload protocol
