+++ localstack-core/localstack/utils/analytics/metrics/__init__.py
@@ -2,13 +2,5 @@
 
 from .counter import Counter, LabeledCounter
 from .registry import MetricRegistry
-from .type import CounterPayload, MetricPayload, MetricRegistryKey
 
-__all__ = [
-    "Counter",
-    "LabeledCounter",
-    "MetricRegistry",
-    "CounterPayload",
-    "MetricPayload",
-    "MetricRegistryKey",
-]
+__all__ = ["Counter", "LabeledCounter", "MetricRegistry"]
